Use checkGZBinaryEqual() to compare BAM files
For BAM files as well as tabix indexes, compare the uncompressed
contents rather than the details of the compression.
Move checkBinaryEqual() and checkGZBinaryEqual() to TestUtils.py.
